Introduce easier naming schema
While programming with OPS, we figured that the current way of method
naming causes a) confusion as it wasn't easily understandable
and b) was pretty inconsistent. Therefore we used the following
naming schema in addition to whats already written down in the wiki:

- RandomAccessibleInterval -> RAI as the only abbreviation
- Mappings are written with "To" instead of "2"
- OpXYInterface -> AbstractOpXY -> DefaultOpXY -> OpXYSpecialFunction
