+++ src/main/java/net/imagej/ops/map/MapConvertRandomAccessToRandomAccess.java
@@ -30,44 +30,30 @@
 
 package net.imagej.ops.map;
 
-import net.imagej.ops.Function;
 import net.imagej.ops.Op;
 import net.imagej.ops.Ops;
-import net.imglib2.Cursor;
-import net.imglib2.IterableInterval;
-import net.imglib2.RandomAccess;
-import net.imglib2.RandomAccessibleInterval;
+import net.imglib2.RandomAccessible;
+import net.imglib2.converter.read.ConvertedRandomAccessible;
+import net.imglib2.type.Type;
 
-import org.scijava.Priority;
 import org.scijava.plugin.Plugin;
 
 /**
- * {@link Map} using a {@link Function} on {@link IterableInterval} and
- * {@link RandomAccessibleInterval}
+ * Maps values of a {@link RandomAccessible} in View.
  * 
- * @author Martin Horn
  * @author Christian Dietz
- * @param <A> mapped on <B>
- * @param <B> mapped from <A>
+ * 
+ * @param <A>
+ * @param <B>
  */
-@Plugin(type = Op.class, name = Ops.Map.NAME, priority = Priority.LOW_PRIORITY)
-public class MapII2RAI<A, B> extends
-	AbstractFunctionMap<A, B, IterableInterval<A>, RandomAccessibleInterval<B>>
+@Plugin(type = Op.class, name = Ops.Map.NAME)
+public class MapConvertRandomAccessToRandomAccess<A, B extends Type<B>> extends
+	MapView<A, B, RandomAccessible<A>, RandomAccessible<B>>
 {
 
 	@Override
-	public RandomAccessibleInterval<B> compute(final IterableInterval<A> input,
-		final RandomAccessibleInterval<B> output)
-	{
-		final Cursor<A> cursor = input.localizingCursor();
-		final RandomAccess<B> rndAccess = output.randomAccess();
-
-		while (cursor.hasNext()) {
-			cursor.fwd();
-			rndAccess.setPosition(cursor);
-			func.compute(cursor.get(), rndAccess.get());
-		}
-
-		return output;
+	public void run() {
+		setOutput(new ConvertedRandomAccessible<A, B>(getInput(), getFunction(),
+			getType()));
 	}
 }
