+++ src/main/java/net/imagej/ops/misc/MinMaxRealType.java
@@ -28,31 +28,47 @@
  * #L%
  */
 
-package net.imagej.ops.map;
+package net.imagej.ops.misc;
+
+import java.util.Iterator;
 
 import net.imagej.ops.Op;
 import net.imagej.ops.Ops;
+import net.imglib2.type.numeric.RealType;
 
-import org.scijava.Priority;
+import org.scijava.ItemIO;
+import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
 
 /**
- * Default (slower) implementation of an {@link MapI}.
- * 
- * @author Curtis Rueden
- * @author Christian Dietz
- * @param <A>
+ * Calculates the minimum and maximum value of an image.
  */
-@Plugin(type = Op.class, name = Ops.Map.NAME,
-	priority = Priority.LOW_PRIORITY)
-public class MapI<A> extends AbstractInplaceMap<A, Iterable<A>> {
+@Plugin(type = Op.class, name = Ops.MinMax.NAME)
+public class MinMaxRealType<T extends RealType<T>> implements MinMax<T> {
+
+	@Parameter
+	private Iterable<T> img;
+
+	@Parameter(type = ItemIO.OUTPUT)
+	private T min;
+
+	@Parameter(type = ItemIO.OUTPUT)
+	private T max;
 
 	@Override
-	public Iterable<A> compute(final Iterable<A> arg) {
-		for (final A t : arg) {
-			func.compute(t, t);
-		}
+	public void run() {
+		min = img.iterator().next().createVariable();
+		max = min.copy();
 
-		return arg;
+		min.setReal(min.getMaxValue());
+		max.setReal(max.getMinValue());
+
+		final Iterator<T> it = img.iterator();
+		while (it.hasNext()) {
+			final T i = it.next();
+			if (min.compareTo(i) > 0) min.set(i);
+			if (max.compareTo(i) < 0) max.set(i);
+		}
 	}
+
 }
