+++ src/test/java/net/imagej/ops/map/view/MapViewTest.java
@@ -34,9 +34,9 @@ import static org.junit.Assert.assertEquals;
 import net.imagej.ops.AbstractOpTest;
 import net.imagej.ops.Op;
 import net.imagej.ops.arithmetic.add.AddConstantToNumericType;
-import net.imagej.ops.map.MapII2View;
-import net.imagej.ops.map.MapRA2View;
-import net.imagej.ops.map.MapRAI2View;
+import net.imagej.ops.map.MapIterableIntervalToView;
+import net.imagej.ops.map.MapConvertRandomAccessToRandomAccess;
+import net.imagej.ops.map.MapConvertRAIToRAI;
 import net.imglib2.Cursor;
 import net.imglib2.IterableInterval;
 import net.imglib2.RandomAccessible;
@@ -68,7 +68,7 @@ public class MapViewTest extends AbstractOpTest {
 	public void testRandomAccessibleView() {
 		@SuppressWarnings("unchecked")
 		final RandomAccessible<ByteType> res =
-			(RandomAccessible<ByteType>) ops.run(MapRA2View.class, in, op,
+			(RandomAccessible<ByteType>) ops.run(MapConvertRandomAccessToRandomAccess.class, in, op,
 				new ByteType());
 
 		final Cursor<ByteType> iterable =
@@ -83,7 +83,7 @@ public class MapViewTest extends AbstractOpTest {
 	public void testRandomAccessibleIntervalView() {
 		@SuppressWarnings("unchecked")
 		final RandomAccessibleInterval<ByteType> res =
-			(RandomAccessibleInterval<ByteType>) ops.run(MapRAI2View.class, in, op,
+			(RandomAccessibleInterval<ByteType>) ops.run(MapConvertRAIToRAI.class, in, op,
 				new ByteType());
 
 		final Cursor<ByteType> iterable = Views.iterable(res).cursor();
@@ -97,7 +97,7 @@ public class MapViewTest extends AbstractOpTest {
 	public void testIterableIntervalView() {
 		@SuppressWarnings("unchecked")
 		final IterableInterval<ByteType> res =
-			(IterableInterval<ByteType>) ops.run(MapII2View.class, in, op,
+			(IterableInterval<ByteType>) ops.run(MapIterableIntervalToView.class, in, op,
 				new ByteType());
 
 		final Cursor<ByteType> iterable = res.cursor();
